This commit is intended to address #198.
We have modified `download_buildpack_cache.rb` so that it looks a
proxy URL on the following variables (in this order):

- HTTPS_PROXY
- https_proxy
- HTTP_PROXY
- http_proxy

If none of these variables is found, the download process is made via `Net::HTTP` as usual.

Else, we extract the proxy host and port from the value of the first variable found (in the specified order).
Then, we substitute the normal `Net::HTTP` with one obtained via `Net::HTTP::Proxy`, configured to use the found
proxy values.