Changed service detection for Dynatrace AppMon integration
The Dynatrace AppMon integration raises an error in case there are more
than one service containing "dynatrace". This might not be supported for
the AppMon integration, but is for the OneAgent integration.

The AppMon integation now only fails if multiple valid services which
are not for the OneAgent integation are bound to the application.